% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNCG.R
\name{checkNegCtrlGenes}
\alias{checkNegCtrlGenes}
\title{Check Unwanted variation in Negative Control Genes}
\usage{
checkNegCtrlGenes(data, ncg_set, group, plot_type, nPcs, npcs, is.log)
}
\arguments{
\item{data}{S4 data object}

\item{ncg_set}{character: A character vector of NCG sets from TCGA data. At present TCGA data includes the following NCG variables i.e. "scRNAseq_HK","RNAseq_HK","Microrray_HK","NanostringPanCancer_HK" and "sinscorePanCancer_HK".}

\item{group}{character: Color code PCs based on group. groups included are 'Time', 'Tissue', 'Plate', 'TSS', 'Center'}

\item{plot_type}{character: Plot type. Includes 'DensityPlot' and 'BoxPlot'.}

\item{nPcs}{numeric: Number of PCs that needs to be computed}

\item{npcs}{numeric: Number of PCs that needs to be plotted}

\item{is.log}{logical: Checks if the S4 data has log values. If 'False', it converts data to log scale.}
}
\value{
Density Plot, Box Plot
}
\description{
This function is a part of the data analysis functionality of \code{tcgaCleaneR}. It helps to visualize the presence
of unwanted variation in the Negative Control Genes (NCG) for RUV-III analysis. Using PCA analysis any relation between
the negative control sets in the TCGA data and the unwanted variation type can be visualized. The plot between PCs and
unwanted variation can help verify the presence of unwanted variation in NCG sets.
}
\examples{
\dontrun{
checkNegCtrlGenes(data =brca.data, ncg_set= c("Microrray_HK"), group='Time', plot_type="DensityPlot", nPcs=10, npcs = 3, is.log=FALSE)
checkNegCtrlGenes(data =brca.data, ncg_set= c("Microrray_HK","scRNAseq_HK"), group='Tissue', plot_type="BoxPlot", nPcs=7, npcs = 2, is.log=FALSE)
}
}
