% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_design.R
\name{plotStudyOutline}
\alias{plotStudyOutline}
\title{Study Design Plot}
\usage{
plotStudyOutline(data)
}
\arguments{
\item{data}{S4 data object: Input TCGA or similar \code{SummarizedExperiment} Dataset}
}
\value{
Heat Map Plot
}
\description{
This function is a part of the data analysis functionality of \code{tcgaCleaneR}. It helps to plot
a Combined Heat Map \code{ComplexHeatmap} for some known unwanted variations and present a dashboard level analysis
of the TCGA data before any analysis.
}
\examples{
\dontrun{
plotStudyOutline(data = brca.data)
}

}
